fix: dedupe OpenAI response parsing (#502) (thanks @ianalloway)
